Small fixes and improvements to the importer
