+++ src/module/dataTemplates.ts
@@ -1,4 +1,11 @@
 import DamageData = Shadowrun.DamageData;
+import TechnologyData = Shadowrun.TechnologyData;
+import DescriptionData = Shadowrun.DescriptionData;
+import EquipmentData = Shadowrun.EquipmentData;
+import QualityData = Shadowrun.QualityData;
+import ActionRollData = Shadowrun.ActionRollData;
+import LimitData = Shadowrun.LimitData;
+import OpposedTestData = Shadowrun.OpposedTestData;
 
 /**
  * TODO: Add unittesting to DefaultValues helper.
@@ -29,6 +36,88 @@ export class DefaultValues {
             mod: [],
             base_formula_operator: 'add'
         }, partialDamageData) as DamageData;
+    }
+
+    static equipmentData(partialEquipmentData: Partial<EquipmentData> = {}): EquipmentData {
+        return mergeObject({
+            description: this.descriptionData(),
+            technology: this.technologyData(),
+        }, partialEquipmentData) as EquipmentData;
+    }
+
+    static qualityData(partialQualityData: Partial<QualityData> = {}): QualityData {
+        return mergeObject({
+            type: '',
+            description: this.descriptionData(),
+            action: this.actionRollData(),
+        }, partialQualityData) as QualityData;
+    }
+
+    static technologyData(partialTechnologyData: Partial<TechnologyData> = {}): TechnologyData {
+        return mergeObject({
+            rating: '',
+            availability: '',
+            quantity: 1,
+            cost: 0,
+            equipped: false,
+            conceal: {
+                base: 0,
+                value: 0,
+                mod: [],
+            },
+            condition_monitor: {
+                label: '',
+                value: 0,
+                max: 0,
+            }
+        }, partialTechnologyData) as TechnologyData;
+    }
+
+    static descriptionData(partialDescriptionData: Partial<DescriptionData> = {}): DescriptionData {
+        return mergeObject({
+            value: '',
+            chat: '',
+            source: ''
+        }, partialDescriptionData) as DescriptionData;
+    }
+
+    static actionRollData(partialActionRollData: Partial<ActionRollData> = {}): ActionRollData {
+        return mergeObject({
+            type: '',
+            category: '',
+            attribute: '',
+            attribute2: '',
+            skill: '',
+            spec: false,
+            mod: 0,
+            mod_description: '',
+            limit: this.limitData(),
+            extended: false,
+            damage: this.damageData(),
+            opposed: this.opposedTestData(),
+            alt_mod: 0,
+            dice_pool_mod: []
+        }, partialActionRollData) as ActionRollData;
+    }
+
+    static limitData(partialLimitData: Partial<LimitData> = {}): LimitData {
+        return mergeObject({
+            value: 0,
+            base: 0,
+            attribute: '',
+            mod: []
+        }, partialLimitData) as LimitData;
+    }
+
+    static opposedTestData(partialOpposedTestData: Partial<OpposedTestData> = {}): OpposedTestData {
+        return mergeObject({
+            type: '',
+            attribute: '',
+            attribute2: '',
+            skill: '',
+            mod: 0,
+            description: ''
+        }, partialOpposedTestData) as OpposedTestData;
     }
 }
