+++ src/module/types/item/Description.ts
@@ -1,9 +1,12 @@
 declare namespace Shadowrun {
+
     export type DescriptionPartData = {
-        description: {
-            value: string;
-            chat: string;
-            source: string;
-        };
+        description: DescriptionData
+    };
+
+    export type DescriptionData = {
+        value: string;
+        chat: string;
+        source: string;
     };
 }
